+++ src/components/sections/about-section.js
@@ -1,7 +1,7 @@
 import React from "react"
-import BackgroundColorLayout from "../../layouts/background-color-layout"
-import TextContainerLayout from "../../layouts/text-container-layout"
-import ExternalLink from "../../external-link"
+import BackgroundColorLayout from "../layouts/background-color-layout"
+import TextContainerLayout from "../layouts/text-container-layout"
+import ExternalLink from "../external-link"
 
 const AboutSection = () => (
   <BackgroundColorLayout
