+++ src/components/templates/blog-template.js
@@ -1,5 +1,6 @@
 import React from "react"
 import { graphql } from "gatsby"
+import BackgroundColorLayout from "../layouts/background-color-layout"
 
 export default function Template({
   data, // this prop will be injected by the GraphQL query below.
@@ -7,16 +8,18 @@ export default function Template({
   const { markdownRemark } = data // data.markdownRemark holds your post data
   const { frontmatter, html } = markdownRemark
   return (
-    <div className="blog-post-container">
-      <div className="blog-post">
-        <h1>{frontmatter.title}</h1>
-        <h2>{frontmatter.date}</h2>
-        <div
-          className="blog-post-content"
-          dangerouslySetInnerHTML={{ __html: html }}
-        />
+    <BackgroundColorLayout>
+      <div className="blog-post-container">
+        <div className="blog-post">
+          <h1>{frontmatter.title}</h1>
+          <h2>{frontmatter.date}</h2>
+          <div
+            className="blog-post-content"
+            dangerouslySetInnerHTML={{ __html: html }}
+          />
+        </div>
       </div>
-    </div>
+    </BackgroundColorLayout>
   )
 }
