+++ src/markdown-pages/post-1.md
@@ -3,3 +3,10 @@ path: "/portfolio/my-first-post"
 date: "2019-05-04"
 title: "My first blog post"
 ---
+
+A small todo-list application using MEAN stack. 
+
+- MongoDB (using mlab)
+- ExpressJS back-end
+- Angular
+- Node
